Add Files and complete Navigation Method
