+++ src/benutzerschnitstelle/Kategorieuebersicht.java
@@ -1,0 +1,6 @@
+package benutzerschnitstelle;
+
+import javax.swing.*;
+
+public class Kategorieuebersicht extends JPanel {
+}
