+++ src/benutzerschnitstelle/VokabelErsteller.java
@@ -1,0 +1,6 @@
+package benutzerschnitstelle;
+
+import javax.swing.*;
+
+public class VokabelErsteller extends JPanel {
+}
