+++ src/steuerung/MainFrameSteuerung.java
@@ -1,10 +1,9 @@
 package steuerung;
 
-import benutzerschnitstelle.Dashboard;
 import benutzerschnitstelle.MainFrame;
-import benutzerschnitstelle.UI;
+import benutzerschnitstelle.UIScreens;
 
-import javax.swing.*;
+import java.awt.*;
 
 public class MainFrameSteuerung {
 
@@ -14,10 +13,42 @@ public class MainFrameSteuerung {
         this.mainFrame = mainFrame;
     }
 
+    /**
+     * Launch the application.
+     */
+    public static void main(String[] args)
+    {
+        EventQueue.invokeLater(() -> {
+            try
+            {
+                MainFrame frame = new MainFrame();
+                frame.setVisible(true);
+            }
+            catch (Exception e)
+            {
+                e.printStackTrace();
+            }
+        });
+    }
+
     public void openDashboard() {
-        mainFrame.open(UI.Dashboard);
+        mainFrame.open(UIScreens.Dashboard);
     }
 
     public void openAbfrage() {
+        mainFrame.open(UIScreens.Abfrage);
+    }
+
+    public void openKategorieuebersicht() {
+        mainFrame.open(UIScreens.Kategorieeuebersicht);
+    }
+
+
+    public void openErsteller() {
+        mainFrame.open(UIScreens.Ersteller);
+    }
+
+    public void openStats() {
+        mainFrame.open(UIScreens.Statistik);
     }
 }
